update opr constants
